refactor(cli): move #999's fused number-canonicalizing walker to yq_runner
Code review (Angle Altitude, corroborated by Angle C/B/Reuse) found the
walker didn't belong in the shared library: it was pub and generic over
any DocumentValue, including YAML's, with nothing stopping a future
YAML or `succinctly jq` call site from reaching for it and silently
regressing #918's number-spelling-preservation guarantee -- the exact
bug class #978/#999 exist to keep away from JSON input specifically.

The pre-existing code (canonicalize_json_numbers, now replaced) kept
this yq-CLI-specific oracle quirk local to yq_runner.rs by design, per
its own doc comment. Restores that: the walker (renamed
to_owned_canonicalizing_numbers) now lives in yq_runner.rs, calling one
small new pub library primitive (OwnedValue::from_number_literal_plain)
for the number arm instead of duplicating the parse-to-scalar mapping.
That primitive also replaces an identical inline match already
duplicated in from_number_bytes's own fallback arm.

Also fixes a real guard mismatch introduced by the walker's move: the
fused single-pass walker needs the same MAX_NESTING_DEPTH (256) guard
to_owned_at_depth uses (this walk has the identical DocumentCursor-based
shape), not MAX_VALUE_TREE_DEPTH (384) -- which the old two-pass code
used only for its second pass over an already-materialized tree, after
the first pass's own 256 guard had already bound the real ceiling.
Reusing the looser 384 guard here would have quietly raised the
effective depth limit for --input-format json documents.

Fixed 5 stale doc-comment references to the deleted
canonicalize_json_numbers name across yq_runner.rs and yaml/light.rs.
